+++ mayan/apps/document_parsing/icons.py
@@ -13,10 +13,6 @@ icon_document_file_content_download = Icon(
     secondary_symbol='arrow-down'
 )
 icon_document_file_page_content = icon_document_file_content
-icon_document_file_parsing_errors_list = Icon(
-    driver_name='fontawesome-dual', primary_symbol='font',
-    secondary_symbol='exclamation'
-)
 icon_document_file_submit_multiple = Icon(
     driver_name='fontawesome-dual', primary_symbol='font',
     secondary_symbol='arrow-right'
@@ -29,10 +25,3 @@ icon_document_type_parsing_settings = Icon(
 # Document type
 
 icon_document_type_submit = icon_document_file_submit_multiple
-
-# Errors
-
-icon_error_list = Icon(
-    driver_name='fontawesome-dual', primary_symbol='font',
-    secondary_symbol='exclamation'
-)