+++ mayan/apps/document_parsing/migrations/0011_delete_documentfileparseerror.py
@@ -1,0 +1,13 @@
+from django.db import migrations
+
+
+class Migration(migrations.Migration):
+    dependencies = [
+        ('document_parsing', '0010_auto_20201012_0708')
+    ]
+
+    operations = [
+        migrations.DeleteModel(
+            name='DocumentFileParseError',
+        )
+    ]