+++ HISTORY.rst
@@ -185,9 +185,12 @@
   - ``ModelReverseField``
   - ``QuerysetParametersSerializer``
 
-- OCR app converted to the new error log system. The permission
+- Convert the OCR app to the new error log system. The permission
   "View error log" is now required to view the document version OCR error
   log.
+- Convert the document parsing app to the new error log system. The
+  permission "View error log" is now required to view the document file
+  parsing error log.
 
 4.2.3 (2022-04-01)
 ==================